concrete person, some comment about uniqueness
