--- core/src/com/jingyuyao/tactical/model/person/BasePerson.java
@@ -1,19 +1,0 @@
-package com.jingyuyao.tactical.model.person;
-
-import com.jingyuyao.tactical.model.resource.ModelBundle;
-import com.jingyuyao.tactical.model.resource.ResourceKey;
-
-class BasePerson implements Person {
-
-  private String name;
-
-  @Override
-  public ResourceKey getName() {
-    return ModelBundle.PERSON_NAME.get(name);
-  }
-
-  @Override
-  public ResourceKey getRole() {
-    return ModelBundle.PERSON_ROLE.get("none");
-  }
-}
